Implement parsing for pointer types
I didn't have to use an Identifier node for the domain type (could just
use a string), but in Extended Pascal, the domain type can be things
other than simple IDs, so I'm using a node so that I have to change less
if I ever get around to implementing that.

Plus it just feels right to always use a node for type references.