Reimplement the requirement that certain tokens must be separated from each other
Now we check for violations of this rule in a more direct way. This follows
the spec more closely, and produces better error messages.

We can now remove the lookahead assertions from most token regexes. However,
the word symbol regexes still require them, because without them, they could
match a part of an identifier.
That's because they fulfill a different role - they are there so that a