Store the map of defining occurrences in Block
The identifier lookup rules require us to be able to check whether an
identifier is defined within a given block, and this map helps us answer
exactly that.

I changed `checkDuplicateIdentifier`'s argument type to `Block&` to make
it easier to use, although it only requires the defining occurrences map.