Fix identifier lookup logic WRT identifiers of wrong types
If we can't find an identifier in the type-specific map, that doesn't
necessarily mean that we need to look in the parent block. The identifier
could be defined in the current block, but belong to a different type
(which is an error). Add this check to the lookup.

As a bonus, implement a dedicated error code/message for when an identifier
is used to define itself.