Make subrange error messages more consistent
The general policy for reporting errors is to report at the
rightmost location where the code could be changed to fix the error.

In both the case where the bound types don't match and the case where
the smallest value is greater than the largest, such rightmost location
is the largest value, so make both errors point there, and rephrase the error
messages to blame the largest value.