Refactor ProgramBuilder::analyzeSubroutineDeclarations
There was some redundancy in the return value of the first `visit` call,
since in the case where a signature was returned, you didn't need `is_function`
to tell whether the subroutine was a function (you could just check whether it
had a return type). Eliminate this redundancy by switching to a variant return
instead.

Also, instead of a boolean, use a custom enum type, to help readability.