Redo the implementation of FieldList
Now all fields (including those of the variant part) are considered fields
of the field list, and their type information is stored in the field list
object. This creates redundancy, but makes it easy to find out what type
a certain field has, which will be useful later.

Fields that are not in the variant part will now be called "own fields".
I'll likely need some accessors for these later.