Change the return type of VariableAccess::type to a reference
I won't be able to return a `shared_ptr` for all expression implementations,
since `Constant::type` returns a reference, and it has to return a reference,
because `ConstantEnumerated` has no `shared_ptr` it can return.

So let's just hope that I never get the object lifetimes confused and return
a reference to something that can go out of scope.