Make Constant::type return a `shared_ptr`
I need to be able, given a subrange type, to construct a set type whose
component type is the host type of the subrange type. This is needed to
implement Pascal's "treated as" rules. For that, the subrange type needs
to be able to return a `shared_ptr` to its host type, and thus the constants
need to be able to return one too.

This requires me to be able to get a `shared_ptr` to `TypeEnumerated` from
a pointer to it, so make it inherit `enable_shared_from_this` and make its
constructor private.

This commit invalidates the reasoning in 0fca5a21, but I'm not going to revert
that just yet, hoping that `Expression::type` will only need to return existing
types.