Majorly refactor the representation for conformant array schemas
There seems to be a lot of code duplication when we deal with array types,
since often we have to replicate similar logic for static arrays and conformant
arrays. Fix this by removing the separate `ConformantArraySchema` class and
instead introducing a `TypeSubrangeDynamic` class that can be used as the index
type for the `TypeArray` class.

The downside of this refactoring is that it's no longer possible to determine
whether a type is a conformant array with one cast. We also can't use polymorphism
to distinguish between static and conformant arrays without casts. This still doesn't
seem to be as bad as the duplication, though.