Prohibit -2^31 from being a valid integer value
Upon re-reading the standard rules, I came to understand that it should not be
allowed. The standard says that values of the integer type must be those
denoted by the `signed-integer` production, which is defined as:

    signed-integer = [sign] unsigned-integer .

And `unsigned-integer` cannot be larger than `maxint`, meaning that the
smallest possible integer is `-maxint`.

This simplifies many things at compile time, although it's going to be annoying
to enforce at run time.